Type fix

+++ kafka-go/pkg/redshift/redshift.go
@@ -707,7 +707,7 @@ var debeziumToRedshiftTypeMap = map[string]string{
 	"int":     "integer",
 	"int16":   "smallint",
 	"int32":   "integer",
-	"long":    "int8",
+	"long":    "bigint",
 	"bigint":  "bigint",
 	"string":  "character varying(256)",
 }
